pyproject.toml
[project]
name = "framework-forge"
version = "0.1.0"
description = "Extract thinking frameworks from historical figures using CDM-adapted methodology"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
include = ["framework_forge*"]

[project.scripts]
framework-forge = "framework_forge.cli:cli"
